requirements.txt
# Core dependencies
feedparser>=6.0.10
python-dotenv>=1.0.0

# Web scraping and parsing
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.8.0
crawl4ai>=0.3.0

# AI and machine learning
openai>=1.0.0
google-generativeai>=0.3.0

# Vector database
pinecone>=3.0.0

# Audio processing
edge-tts>=6.1.0
pygame>=2.5.0

# Date/time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Image processing
pillow>=10.0.0

# Scheduling and background tasks
schedule>=1.2.0

# Utilities
pydantic>=2.0.0
numpy>=1.24.0
python-magic>=0.4.27
protobuf~=4.25.2
scikit-learn~=1.4.1.post1
streamlit~=1.28.0